add StartMenu en-US translations
