--- assets/locales/en-US/hello_world.ftl
@@ -1,1 +1,0 @@
-hello-world = hello world
