+++ assets/locales/en-US/main.ftl.yml
@@ -1,4 +1,4 @@
 # Locale files may be in YAML as in this example, or in RON
 locale: en-US
 resources:
-  - hello_world.ftl
+  - start_menu.ftl
