+++ assets/locales/en-US/start_menu.ftl
@@ -1,0 +1,2 @@
+press-start = Press Start
+credits-prompt = Press 'C' for Credits
