+++ src/main.rs
@@ -5,7 +5,6 @@ use bevy_asset_loader::prelude::*;
 use {bevy_inspector_egui::quick::WorldInspectorPlugin, bevy_rapier2d::prelude::*};
 
 mod core;
-mod i18n;
 mod inputs;
 mod shaders;
 mod ships;
@@ -16,11 +15,11 @@ mod world;
 
 use crate::{
     core::resources::{
-        assets::{AudioAssets, DataAssets, SpriteAssets, UiAssets},
+        assets::{AudioAssets, DataAssets, I18nAssets, SpriteAssets, UiAssets},
         config::ConfigPlugin,
+        i18n::I18nPlugin,
     },
     core::CorePlugin,
-    i18n::I18nPlugin,
     ships::ShipsPlugin,
     systems::{events::EventsPlugin, states::GameState, SystemsPlugin},
     ui::UiPlugin,
@@ -69,13 +68,16 @@ fn main() {
         WorldInspectorPlugin::new(),
     ));
 
+    // TODO: Assets should be separated according to where and when they are needed.
+    //       Right now, we're loading ALL assets at startup. This is unnecessary.
     app.add_loading_state(
         LoadingState::new(GameState::Loading).continue_to_state(GameState::StartMenu),
     )
     .add_collection_to_loading_state::<_, SpriteAssets>(GameState::Loading)
     .add_collection_to_loading_state::<_, AudioAssets>(GameState::Loading)
     .add_collection_to_loading_state::<_, UiAssets>(GameState::Loading)
-    .add_collection_to_loading_state::<_, DataAssets>(GameState::Loading);
+    .add_collection_to_loading_state::<_, DataAssets>(GameState::Loading)
+    .add_collection_to_loading_state::<_, I18nAssets>(GameState::Loading);
 
     app.insert_resource(ClearColor(Color::rgb(0., 0., 0.)));
